Add developer mode to one set of credentials The credentials: Username: Password:
In developer mode, when they enter a category (like e-commerce or smth), there should be a button on the top right to "add codes".

The user should then be directed to the following form:

(Site for normal mode, Game for gaming mode):
Code:
Expiry date:
Description:

Once the form is filled, this should be added to the set of codes visible in that category.