Add a new feature to developer mode that allows developers to add PromoPoints to a contributor on the Top Contributors Leaderboard.
Feature Description:
When a user is in developer mode, display an "Add PromoPoints" button on the Top Contributors Leaderboard page.

Clicking the button should open an in-site form (no page reload) with the following fields:

-> Contributor Username (text input)

-> Number of PromoPoints Gained (numeric input)

On form submission:

-> Validate the inputs

-> Find the contributor by username

-> Add the entered number of PromoPoints to their existing total

-> Save the updated score in the leaderboard's data store (e.g., JSON file, database, etc.)

Only show this form and functionality to users who are in developer mode.